ws/requirements.txt
fastapi
uvicorn
websockets
motor
apscheduler
python-dotenv